+++ main.go
@@ -5,12 +5,11 @@ package main
 import (
 	"context"
 	"fmt"
-	"log"
 	"log/slog"
 	"os"
 
 	"github.com/urfave/cli/v2"
-	"github.com/wayneashleyberry/gh-arc/pkg/cmd"
+	"github.com/wayneashleyberry/gh-arc/pkg/gomod"
 )
 
 func setDefaultLogger(level slog.Leveler) {
@@ -27,7 +26,8 @@ func main() {
 	ctx := context.Background()
 
 	if err := run(ctx); err != nil {
-		log.Fatal(err)
+		fmt.Println(err)
+		os.Exit(1)
 	}
 }
 
@@ -64,7 +64,7 @@ func run(_ context.Context) error {
 				Action: func(c *cli.Context) error {
 					checkIndirect := c.Bool("indirect")
 
-					count, err := cmd.ListArchivedGoModules(c.Context, checkIndirect)
+					count, err := gomod.ListArchived(c.Context, checkIndirect)
 					if err != nil {
 						return fmt.Errorf("failed to list archived go modules: %w", err)
 					}
